submission-args: tests
